add feature checker selecting one broadcast and updating Item and PO details, make modular schema
